Use default Composer type for Drupal modules
Modules on Drupal.org uses the drupal-module Composer type. This change
ensures that this project does the same.

Consequently there is no need have a dependency to composer/installers
either.

If someone wants to treat this module - or all modules from a specific
vendor differently they then can add and use composer/installers to do
so.