Allow/prevent bot crawling based on settings.php
This allows you to keep the module enabled on all sites but control
if crawling should be allowed based on settings.php.

This brings the Drupal 8 version up to feature parity with the Drupal 7
version.

Update README accordingly. The only difference is that the configuration
goes in $settings - not in $conf.